refactor(rta): extract RTA functional-test layer (tests/rta/) shared with screenshots
Sets RTA up properly. The reusable device driver, nav/wait steps, registry
seeds, Jellyfin demo helpers, and a single screen registry now live under
tests/rta/ and are shared by (a) Vitest functional tests that assert each
screen loads and (b) the store-screenshot generator.

- tests/rta/: lib/{driver,steps,seed,jellyfin,nav}, screens.js (registry),
  capture.js (RTA_CAPTURE raw UI), specs/screens.spec.js, vitest.rta.config.js.
- npm: test:rta / :tdd / :fast / :capture (test:rta passes 6/6 screens on device).
- Adds an explicit userSelect loaded-assertion (previously none).
- scripts/capture-screenshots.js is now a thin orchestrator importing the shared
  lib + adding only the store layer (locale matrix, ffmpeg backdrop, manifest).
- Store screenshots default to the PROD build (release branding); :dev / :fast
  variants. Verified prod keeps the ENABLE_RTA passthrough (RTA-compatible).
- Docs: docs/dev/rta-tests.md, tests/rta/CLAUDE.md; updated tests/CLAUDE.md,
  root CLAUDE.md, architecture/testing.md; removed the tests/source/e2e/
  placeholder (resolves e2e-folder-empty tech-debt); added
  weblate-branch-stale-sync tech-debt.

Refs #621.
