+++ dictionary.txt
@@ -318,6 +318,7 @@ NTSC
 Nygard
 Nygard's
 onboarding
+ODC
 OpenAPI
 opencode
 orchestrator
@@ -440,6 +441,7 @@ SHA-1
 ship-today
 sideload
 Sideload
+sideloaded
 sideloads
 signals-schema-invalid
 slow-decay
