+++ package.json
@@ -113,8 +113,14 @@
     "test:complete": "npm run build:tests-complete && node scripts/run-roku-tests.js",
     "test:scripts": "vitest run",
     "test:scripts:tdd": "vitest",
+    "test:rta": "npm run build && vitest run --config vitest.rta.config.js",
+    "test:rta:tdd": "vitest --config vitest.rta.config.js",
+    "test:rta:fast": "RTA_NO_DEPLOY=1 vitest run --config vitest.rta.config.js",
+    "test:rta:capture": "npm run build && RTA_CAPTURE=1 vitest run --config vitest.rta.config.js",
     "prepare": "husky",
-    "screenshots:capture": "node scripts/capture-screenshots.js"
+    "screenshots:capture": "npm run build:prod && DEPLOY=1 node scripts/capture-screenshots.js",
+    "screenshots:capture:dev": "npm run build && DEPLOY=1 node scripts/capture-screenshots.js",
+    "screenshots:capture:fast": "node scripts/capture-screenshots.js"
   },
   "repository": {
     "type": "git",
