--- tests/source/e2e/README.md
@@ -1,27 +1,0 @@
-# End-to-End Tests
-
-**Purpose:** Test complete user journeys through the application UI.
-
-**Status:** 🚧 Placeholder for future implementation
-
-**Planned Tool:** [RTA (Roku Test Automation)](https://github.com/triwav/roku-test-automation)
-
-**When to use:**
-
-- Testing critical user flows (login, playback, navigation)
-- Testing across different Roku device models
-- Regression testing before major releases
-- UI automation with remote control simulation
-
-**Characteristics:**
-
-- Requires actual Roku device
-- Slowest execution (minutes)
-- High confidence in real-world behavior
-- Run before releases or nightly
-
-**Future Implementation:**
-
-- Add RTA as dev dependency
-- Create test scenarios for critical paths
-- Integrate with CI/CD for release builds
